requirements.txt
fastapi>=0.115.6,<1.0
uvicorn[standard]>=0.32.1,<1.0
pydantic>=2.10.4,<3.0
httpx>=0.28.1,<1.0
google-genai>=1.0.0,<2.0
google-cloud-translate>=3.15.0,<4.0
google-cloud-texttospeech>=2.16.0,<3.0
google-cloud-bigquery>=3.27.0,<4.0
google-cloud-dlp>=3.25.0,<4.0
